+++ accent-common/src/test/java/de/embl/rieslab/accent/common/data/calibration/CalibrationIOTest.java
@@ -19,7 +19,7 @@ public class CalibrationIOTest {
 		Calibration c2 = generateCalibration();
 		assertTrue(Calibration.areEquals(c1, c2));
 		
-		double[] gain =  { 30.1, 14.1, 25.1, 34.1 };
+		float[] gain =  { 30.1f, 14.1f, 25.1f, 34.1f };
 		c2.setGain(gain);
 		assertFalse(Calibration.areEquals(c1, c2));
 	}
@@ -51,7 +51,7 @@ public class CalibrationIOTest {
 	
 	@Test 
 	public void testWrongSizeException() {
-		double[] d = { 1, 2, 3 };
+		float[] d = { 1, 2, 3 };
 		Calibration c = generateCalibration();
 		
 		assertThrows(IllegalArgumentException.class, () -> {
@@ -90,15 +90,15 @@ public class CalibrationIOTest {
 	
 	public static Calibration generateCalibration() {
 		// Creates calibration object
-		double[] baseline = { 0.1, 1.1, 2.1, 3.1 };
-		double[] dc_per_sec = { 4.1, 5.1, 6.1, 7.1 };
-		double[] r_sq_avg = { 8.1, 9.1, 10.1, 11.1 };
-		double[] rn_sq = { 12.1, 13.1, 14.1, 15.1 };
-		double[] tn_sq_per_sec = { 10.1, 11.1, 21.1, 31.1 };
-		double[] r_sq_var = { 20.1, 12.1, 22.1, 32.1 };
+		float[] baseline = { 0.1f, 1.1f, 2.1f, 3.1f };
+		float[] dc_per_sec = { 4.1f, 5.1f, 6.1f, 7.1f };
+		float[] r_sq_avg = { 8.1f, 9.1f, 10.1f, 11.1f };
+		float[] rn_sq = { 12.1f, 13.1f, 14.1f, 15.1f };
+		float[] tn_sq_per_sec = { 10.1f, 11.1f, 21.1f, 31.1f };
+		float[] r_sq_var = { 20.1f, 12.1f, 22.1f, 32.1f };
 
-		double[] gain = { 30.1, 14.1, 24.1, 34.1 };
-		double[] r_sq_gain = { 50.1, 15.1, 25.1, 35.1 };
+		float[] gain = { 30.1f, 14.1f, 24.1f, 34.1f };
+		float[] r_sq_gain = { 50.1f, 15.1f, 25.1f, 35.1f };
 
 		return new Calibration(2, 2, baseline, dc_per_sec, r_sq_avg, 
 				rn_sq, tn_sq_per_sec, r_sq_var, gain, r_sq_gain);
